app/requirements.txt
requests==2.31.0
pandas==2.1.1
playwright==1.50.0
tf-playwright-stealth==1.1.1
streamlit==1.42.0
altair==5.0.1
matplotlib==3.8.0
seaborn==0.12.2
python-docx==0.8.11
openai==0.28.0
price-parser==0.3.4
sib-api-v3-sdk==7.6.0
greenlet==3.1.1
plotly==6.0.0
groq==0.18.0
streamlit-shadcn-ui==0.1.18
twilio==8.12.0
psycopg2-binary==2.9.10
supabase==1.29.0
authlib>=1.3.2
pgvector==0.3.6
